+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 - SlashContext.edit_last_response can now be used to edit a deferred response.
+- Small typing fixes made while setting pyright to strict.
 
 ### Removed
 - suppress_exceptions from Client.dispatch_client_callback cause it was half-arsed and didn't make sense.